Add seek support to ConcatenatedReadStream
Enable random access in ConcatenatedReadStream by implementing seek logic and position tracking. Update PipelineContext to use long for chunk lengths, supporting large files. Refactor read logic for both sequential and seekable access. Adjust FileController to provide chunk lengths as long values. Remove disposal of hashes enumerator in favor of List-based chunk management.

+++ src/Cotton.Storage/Pipelines/PipelineContext.cs
@@ -4,6 +4,6 @@
     {
         public long? FileSizeBytes { get; set; }
         public bool StoreInMemoryCache { get; set; }
-        public Dictionary<string, int>? ChunkLengths { get; set; }
+        public Dictionary<string, long>? ChunkLengths { get; set; }
     }
 }
